# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "kokodah222@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/stripe": {
            "get": {
                "tags": [
                    "Stripe"
                ],
                "summary": "Get list of user stripe checkout session and subscription",
                "description": "get list of user stripe checkout session and subscriptions",
                "operationId": "getUserStripeCheckoutSessionAndSubscription",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "User not found"
                    }
                }
            }
        },
...
        "/stripe/add/product": {
            "post": {
                "tags": [
                    "Stripe"
                ],
                "summary": "Create new stripe product",
                "description": "Create new stripe product",
                "operationId": "StripeAddProduct",
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/stripe/add/customer/{id}": {
            "post": {
                "tags": [
                    "Stripe"
                ],
                "summary": "Create new stripe customer",
                "description": "Create new stripe customer",
                "operationId": "StripeAddCustomer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/stripe/products/{id}": {
            "get": {
                "tags": [
                    "Stripe"
                ],
                "summary": "Get list of user stripe products",
                "description": "Returns list of user stripe products",
                "operationId": "getUserStripeProducts",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/user/subscriptions": {
            "get": {
                "tags": [
                    "Stripe"
                ],
                "summary": "Get list of user stripe subscriptions",
                "description": "Returns list of user stripe subscripions",
                "operationId": "getUserStripeSubscription",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get list of user",
                "description": "Returns list of user",
                "operationId": "getUser",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "User not found"
                    }
                }
            }
        },
...
    },
...
}